rename src/main classes and clean up fig-gen CLI
Renamed main classes for build.gradle file. It makes sense to shift the
generic class name to the CLI class because the CLI calls the GUI and
ismore of a root now.

Figure_Generation CLI classes now have options display in an unsorted
order, rather than alphabetical. This will keep related options grouped
together in the display of help messages.

Figure_Generation CLI classes have validateInput methods that now
return one big string rather than Integers. This will allow for the
printing of all problematic user intputs together, rather than having
the user iterate between fixing an option and then rerunning the whole
program to find another input was also problematic.

Figure_Generation CLI classes had unnesscary  imports removed,
whitespace was cleaned up, and more comments were added. These comments
include comments that are placeholders for future updates/fixes to add
validation or enhancements to the tool.
